requirements.txt
fastapi
uvicorn
numpy
mysql-connector-python
python-multipart
